refactor: move all MapStorage code into a separate file
